Fix support for gret 1.0.0+
Blender requires that the argument passed to a BoolVectorProperty has length matching the BoolVectorProperty, so when there's less modifiers than that length, the modifier_mask argument has to be padded.

Additionally, in gret 1.0.0, the context pointer used changed (only in the execute method) to .active_object, so that has to be overridden alongside .object otherwise the modifiers are applied to the wrong object.

Should the length of the modifier_mask change in the future, it's length is now retrieved dynamically instead of being hardcoded.